+++ lib/flexi_signal_model.dart
@@ -1,6 +1,7 @@
 class InvokeModel {
   String? methodName;
   List<Object>? arguments;
+  String? listenMethodName;
 
-  InvokeModel({this.methodName, this.arguments});
+  InvokeModel({this.methodName, this.arguments, this.listenMethodName});
 }