+++ CHANGELOG.md
@@ -2,3 +2,8 @@
 
 - SignalR library integrated
 - FlexiSignalR class declared
+
+## 0.0.2
+
+- FlexiSignal method modified
+- Additional methods added